+++ CHANGELOG.md
@@ -1,8 +1,16 @@
 # Changelog
 
+
+## [unreleased]
+
+* Added a `modMail` handler to send logs via emails using MODX email service
+* The service now complies with `LoggerInterface` so we could `$service->error()` or `$modx->logger->error()`
+
+
 ## [v0.1.0] - 2016-09-12
 
 * First build
 
 
+[unreleased]: https://github.com/meltingmedia/monolog/compare/0.1.0...HEAD
 [v0.1.0]: https://github.com/meltingmedia/monolog/commits/0.1.0